address comments from jeff
