+++ src/whitenoise/streaming.rs
@@ -104,7 +104,7 @@ impl Whitenoise {
         })
     }
 
-    pub(crate) fn drain_user_updates(
+    fn drain_user_updates(
         mut initial_user: User,
         updates: &mut broadcast::Receiver<user_streaming::UserUpdate>,
     ) -> Result<User> {
@@ -257,5 +257,45 @@ impl Whitenoise {
     /// (filesystem) and database layer (metadata) for media files.
     pub(crate) fn media_files(&self) -> media_files::MediaFiles<'_> {
         media_files::MediaFiles::new(&self.storage, &self.database)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use chrono::Utc;
+    use nostr_sdk::{Keys, Metadata};
+    use tokio::sync::broadcast;
+
+    use super::*;
+    use crate::whitenoise::user_streaming;
+
+    #[test]
+    fn test_drain_user_updates_uses_newest_queued_update() {
+        let pubkey = Keys::generate().public_key();
+        let mut initial_user = User {
+            id: None,
+            pubkey,
+            metadata: Metadata::new().name("Old Name"),
+            created_at: Utc::now(),
+            metadata_known_at: None,
+            updated_at: Utc::now(),
+        };
+        initial_user.mark_metadata_known_now();
+
+        let mut newest_user = initial_user.clone();
+        newest_user.metadata = Metadata::new().name("Newest Name");
+        newest_user.updated_at = Utc::now();
+
+        let (sender, mut updates) = broadcast::channel(10);
+        sender
+            .send(user_streaming::UserUpdate {
+                trigger: user_streaming::UserUpdateTrigger::MetadataChanged,
+                user: newest_user.clone(),
+            })
+            .expect("should queue update");
+
+        let drained_user = Whitenoise::drain_user_updates(initial_user, &mut updates).unwrap();
+
+        assert_eq!(drained_user, newest_user);
     }
 }
